+++ src/api/hooks/Card/GetCardInfo.ts
@@ -1,12 +1,12 @@
 import { useQuery } from '@tanstack/react-query';
 import { keys } from '../../utils/createQueryKey';
-import instnace from '../../../axios/api';
+import apis from '../../axios/api';
 
 export const GetCardInfo = () => {
   const { data, isLoading } = useQuery({
     queryKey: [keys.GET_CARD_INFO],
     queryFn: async () => {
-      const response = await instnace.get('/usersInfo');
+      const response = await apis.get('/usersInfo');
       return response.data.user;
     },
   });
