+++ src/api/hooks/Feed/useCheckTodo.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import instnace from '../../../axios/api';
+import apis from '../../axios/api';
 import { keys } from '../../utils/createQueryKey';
 
 export const useCheckTodo = () => {
@@ -7,7 +7,7 @@ export const useCheckTodo = () => {
 
   const { mutate } = useMutation({
     mutationFn: async (todoId: number) => {
-      const data = await instnace.patch(`/feed/todo/${todoId}`);
+      const data = await apis.patch(`/feed/todo/${todoId}`);
       return data.data;
     },
     onSuccess: () => {
