+++ src/api/hooks/Feed/useDeleteTodo.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import instnace from '../../../axios/api';
+import apis from '../../axios/api';
 import { keys } from '../../utils/createQueryKey';
 
 export const useDeleteTodo = () => {
@@ -7,7 +7,7 @@ export const useDeleteTodo = () => {
 
   const { mutate } = useMutation({
     mutationFn: async (todoId: number) => {
-      const data = await instnace.delete(`feed/todo/${todoId}`);
+      const data = await apis.delete(`feed/todo/${todoId}`);
       return data.data;
     },
     onSuccess: () => {
