+++ src/api/hooks/Feed/usePostTodo.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import instnace from '../../../axios/api';
+import apis from '../../axios/api';
 import { keys } from '../../utils/createQueryKey';
 import { AxiosError, isAxiosError } from 'axios';
 
@@ -15,7 +15,7 @@ export const usePostTodo = () => {
 
   const { mutate } = useMutation({
     mutationFn: async (payload: TodoPayload) => {
-      const data = await instnace.post(
+      const data = await apis.post(
         `/feed/category/${payload.categoryId}/todo`,
         payload.content
       );
