+++ src/api/hooks/Main/usePostschedule.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import api from '../../../axios/api';
+import api from '../../axios/api';
 import React from 'react';
 import { keys } from '../../utils/createQueryKey';
