+++ src/api/hooks/Request.tsx/useGetRequestDetail.ts
@@ -1,12 +1,12 @@
 import { useQueryClient } from '@tanstack/react-query';
 import { keys } from '../../utils/createQueryKey';
-import instnace from '../../../axios/api';
+import apis from '../../axios/api';
 
 export const useGetRequestDetail = () => {
   const queryClient = useQueryClient();
 
   const detail = async (id: number) => {
-    const response = await instnace.get(`/schedule/${id}`);
+    const response = await apis.get(`/schedule/${id}`);
     queryClient.setQueryData([keys.GET_REQUEST_DETAIL], response.data);
   };
