+++ src/api/hooks/Tag/useGetTag.ts
@@ -1,5 +1,5 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
-import instnace from '../../../axios/api';
+import apis from '../../axios/api';
 import { keys } from '../../utils/createQueryKey';
 import { AxiosError } from 'axios';
 
@@ -27,7 +27,7 @@ export const useMentionedSchedules = () => {
   >({
     queryKey: [keys.GET_TAG], // 쿼리 키
     queryFn: async ({ pageParam = 1 }) => {
-      const response = await instnace.get(
+      const response = await apis.get(
         `/mentionedSchedule?pageSize=25&pageNum=${pageParam}`
       );
       return { ...response.data, pageNum: pageParam };
