+++ src/api/hooks/Tag/usePatchTag.ts
@@ -1,12 +1,12 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import instnace from '../../../axios/api';
+import apis from '../../axios/api';
 import { keys } from '../../utils/createQueryKey';
 
 export const usePatchTag = () => {
   const queryClient = useQueryClient();
   const { mutate } = useMutation({
     mutationFn: async (mentionId: number) => {
-      const data = await instnace.patch(`/mentionedSchedule/${mentionId}`);
+      const data = await apis.patch(`/mentionedSchedule/${mentionId}`);
       return data.data;
     },
     onSuccess: () => {
