+++ src/api/hooks/Vacation/usePutDecision.ts
@@ -1,5 +1,5 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
-import instnace from '../../../axios/api';
+import apis from '../../axios/api';
 import { keys } from '../../utils/createQueryKey';
 
 interface Payload {
@@ -11,7 +11,7 @@ export const usePutDecision = () => {
   const queryClient = useQueryClient();
   const { mutate } = useMutation({
     mutationFn: async (payload: Payload) => {
-      const result = await instnace.put(`/vacation/${payload.eventId}/${payload.status}`);
+      const result = await apis.put(`/vacation/${payload.eventId}/${payload.status}`);
       return result.data;
     },
     onSuccess: () => {
