+++ src/pages/MasterSignup/hooks/useCompanyIdValidation.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useMutation } from '@tanstack/react-query';
-import instnace from '../../../axios/api';
+import apis from '../../../api/axios/api';
 
 export const useCompanyIdValidation = () => {
   const [companyIdValidation, setCompanyValidation] = React.useState<boolean>(false);
@@ -10,7 +10,7 @@ export const useCompanyIdValidation = () => {
 
   const checkCompanyId = useMutation(
     async (item: string) => {
-      const data = await instnace.post(`/idCheck`, {
+      const data = await apis.post(`/idCheck`, {
         companyId: item,
       });
       return data;
